+++ test.sh
@@ -1,0 +1,3 @@
+for run_test in ./test_*.sh; do
+  $run_test 2> /dev/null
+done